# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: Ihyaa REST API — واجهة برمجة تطبيقات إحياء
  version: 1.0.0
  description: |
    # Ihyaa (إحياء) — REST API

    منصة إحياء تربط أصحاب الأفكار بالمستثمرين عبر تقييم ذكي بالذكاء الاصطناعي.
    Ihyaa connects idea owners with investors through AI-powered project evaluation.

    ## المصادقة / Authentication
    تسجيل الدخول يمنح توكن Sanctum صالحاً لمدة 24 ساعة يُرسل في الرأس:
    `Authorization: Bearer {token}` — `SRS-NFR-07`.
    Login issues a 24-hour Sanctum token sent via the header above.

    ## الأخطاء / Errors
    جميع الأخطاء تتبع الهيكل الموحد `{code, message, errors}` —
    مع رؤوس `X-RateLimit-*` في كل استجابة و`Retry-After` عند 429 (انظر `rate-limiting-spec.md`).

    ## المصادر المرجعية / References
    - SRS-MVP v3 §5.4.5 (SRS-API-01..49) · `requirements/srs-mvp-v3.md`
    - Rate Limiting: `requirements/rate-limiting-spec.md` (7 مستويات L1..L7)
    - القيم الثابتة: `docs/api/enums.md`

  contact:
    name: Ihyaa Engineering Team
  license:
    name: Proprietary — MVP internal
    identifier: LicenseRef-proprietary
servers:
  - url: https://api.ihyaa.app/api
    description: الإنتاج / Production
  - url: http://127.0.0.1:8000/api
    description: التطوير المحلي / Local development
security:
  - {}
...
paths:
...
  /email/verify:
    post:
      tags: [Auth]
      summary: تفعيل البريد برمز OTP / Verify email (SRS-API-04)
      description: |
        تحقق من رمز OTP (6 أرقام، صالح دقيقة واحدة). عند حذف `code` يُعاد إرسال رمز جديد.
        يمنح التوكن عند التحقق الناجح. Rate Limit: 3/دقيقة لكل email (auth.email-verify).
      operationId: verifyEmail
      x-srs-api: SRS-API-04
      x-rate-limit: "3/دقيقة · email · L1"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email]
              properties:
                email: { type: string, format: email }
                code: { type: string, description: رمز OTP من 6 أرقام — يُحذف لإعادة الإرسال / 6-digit OTP; omit to resend }
            examples:
              verify:
                summary: تحقق بالرمز / Verify with code
                value: { email: ahmed@example.com, code: "483920" }
              resend:
                summary: إعادة إرسال / Resend
                value: { email: ahmed@example.com }
      responses:
        "200":
          description: تحقق ناجح (توكن) أو إعادة إرسال / Verified (token) or resent
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AuthResponse" }
        "409":
          description: البريد مفعّل مسبقاً / Already verified
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ApiError" }
              example: { code: EMAIL_ALREADY_VERIFIED, message: البريد مفعّل مسبقاً }
        "422":
          description: رمز غير صحيح أو منتهي / Invalid or expired OTP
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ApiError" }
              examples:
                invalid: { summary: رمز غير صحيح, value: { code: OTP_INVALID, message: الرمز غير صحيح } }
                expired: { summary: رمز منتهي, value: { code: OTP_EXPIRED, message: انتهت صلاحية الرمز. اطلب رمزاً جديداً } }
        "429":
          $ref: "#/components/responses/RateLimitError"
  /forgot-password:
    post:
      tags: [Auth]
      summary: طلب إعادة تعيين كلمة المرور / Forgot password (SRS-API-05)
      description: |
        يرسل رابط إعادة تعيين صالحاً لمدة ساعة — يعيد 200 دائماً لمنع تسريب وجود البريد.
        Rate Limit: 2/دقيقة لكل email (auth.forgot-password).
      operationId: forgotPassword
      x-srs-api: SRS-API-05
      x-rate-limit: "2/دقيقة · email · L1"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email]
              properties:
                email: { type: string, format: email }
            example: { email: ahmed@example.com }
      responses:
        "200":
          description: أُرسل الرابط إن وُجد الحساب / Link sent if account exists
          content:
            application/json:
              schema: { $ref: "#/components/schemas/MessageResponse" }
              example: { message: إذا كان البريد مسجلاً في المنصة، ستصل رسالة إعادة التعيين خلال دقائق (الرابط صالح لمدة ساعة) }
        "422":
          $ref: "#/components/responses/ValidationError"
        "429":
          $ref: "#/components/responses/RateLimitError"
  /reset-password:
    post:
      tags: [Auth]
      summary: إعادة تعيين كلمة المرور / Reset password (SRS-API-06)
      description: |
        تعيين كلمة مرور جديدة بالرمز المستلم عبر البريد. الرابط صالح ساعة واحدة.
        Rate Limit: 2/دقيقة لكل email (auth.reset-password).
      operationId: resetPassword
      x-srs-api: SRS-API-06
      x-rate-limit: "2/دقيقة · email · L1"
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ResetPasswordRequest" }
            example:
              email: ahmed@example.com
              token: 8e2f9c1a4b7d
              password: "New!Pass123"
              password_confirmation: "New!Pass123"
      responses:
        "200":
          description: تم التحديث / Password updated
          content:
            application/json:
              schema: { $ref: "#/components/schemas/MessageResponse" }
              example: { message: تم تحديث كلمة المرور بنجاح. يمكنك تسجيل الدخول الآن }
        "422":
          description: رابط غير صالح أو تحقق فاشل / Invalid link or validation
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ApiError" }
              example: { code: RESET_LINK_INVALID, message: الرابط غير صالح أو منتهي الصلاحية }
        "429":
          $ref: "#/components/responses/RateLimitError"
...
  /projects/{project}/evaluation-status:
    get:
      tags: [Evaluation]
      summary: حالة التقييم / Evaluation status (SRS-API-47)
      description: |
        استعلام حالة التقييم الحالي (processing/completed/failed) — للمالك.
        يُستخدم مع WebSocket لتحديث الواجهة. Rate Limit: shared.read.
      operationId: evaluationStatus
      x-srs-api: SRS-API-47
      security: [{ bearerAuth: [] }]
      parameters:
        - $ref: "#/components/parameters/ProjectPath"
      responses:
        "200":
          description: الحالة الحالية / Current status
          content:
            application/json:
              schema: { $ref: "#/components/schemas/EvaluationStatusResponse" }
        "401":
          $ref: "#/components/responses/Unauthenticated"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "429":
          $ref: "#/components/responses/RateLimitError"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: Sanctum Token
      description: توكن Sanctum صالح 24 ساعة — يحصل عليه من /login أو /email/verify أو OAuth callback
  parameters:
...
    ProjectPath:
      name: project
      in: path
      required: true
      description: معرّف المشروع / Project ID
      schema: { type: integer, format: int64 }
...
  responses:
    Unauthenticated:
      description: توكن مفقود أو منتهي / Missing or expired token
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ApiError" }
          example: { code: UNAUTHENTICATED, message: يجب تسجيل الدخول للوصول إلى هذا المورد }
    Forbidden:
      description: لا صلاحية — دور أو ملكية أو إفصاح / Forbidden — role, ownership or disclosure
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ApiError" }
          example: { code: FORBIDDEN, message: ليس لديك صلاحية الوصول إلى هذه الصفحة }
    NotFound:
      description: المورد غير موجود / Resource not found
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ApiError" }
          example: { code: NOT_FOUND, message: المورد المطلوب غير موجود }
...
    ValidationError:
      description: فشل التحقق من صحة البيانات / Validation failed
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ValidationError" }
          example:
            code: VALIDATION_ERROR
            message: البيانات المدخلة غير صحيحة
            errors:
              description: [ يجب أن يتراوح طول الوصف بين 50 و2000 حرف ]
    RateLimitError:
      description: تجاوز معدل الطلبات — مع رؤوس X-RateLimit-* وRetry-After / Rate limit exceeded
      headers:
        X-RateLimit-Limit: { schema: { type: integer } }
        X-RateLimit-Remaining: { schema: { type: integer } }
        X-RateLimit-Reset: { schema: { type: integer } }
        Retry-After: { schema: { type: integer } }
      content:
        application/json:
          schema: { $ref: "#/components/schemas/RateLimitError" }
...
  schemas:
...
    UserRole:
      type: string
      enum: [idea_owner, investor, admin]
      description: دور المستخدم — admin عبر seeder فقط / Role — admin via seeder only
...
    EvaluationStatus:
      type: string
      enum: [processing, completed, failed, partial]
      description: حالة تقييم AI / Evaluation status
...
    ApiError:
      type: object
      required: [code, message]
      properties:
        code: { type: string, description: رمز الخطأ الآلي / Machine-readable code }
        message: { type: string, description: رسالة مترجمة / Localized message }
        errors:
          type: object
          additionalProperties:
            type: array
            items: { type: string }
          description: تفاصيل تحقق لكل حقل / Per-field validation details
    ValidationError:
      allOf:
        - { $ref: "#/components/schemas/ApiError" }
      required: [code, errors]
      properties:
        code: { type: string, const: VALIDATION_ERROR }
    RateLimitError:
      type: object
      required: [code, message, retry_after, reset_at]
      properties:
        code: { type: string, const: RATE_LIMIT_EXCEEDED }
        message: { type: string }
        retry_after: { type: integer, description: ثوانٍ حتى المحاولة التالية / Seconds until retry }
        reset_at: { type: integer, description: Unix timestamp لإعادة التعيين / Reset timestamp }
    MessageResponse:
      type: object
      required: [message]
      properties:
        message: { type: string }
...
    User:
      type: object
      required: [id, name, email, role]
      properties:
        id: { type: integer, format: int64 }
        name: { type: string }
        email: { type: string, format: email }
        role: { $ref: "#/components/schemas/UserRole" }
        avatar_url: { type: [string, "null"], format: uri }
        email_verified_at: { type: [string, "null"], format: date-time }
        provider: { type: [string, "null"], enum: [google, github, linkedin] }
        created_at: { type: string, format: date-time }
    AuthResponse:
      type: object
      required: [message, token, user]
      properties:
        message: { type: string }
        token: { type: string, description: توكن Sanctum — 24 ساعة / Sanctum token — 24h }
        token_type: { type: string, const: Bearer }
        expires_in: { type: integer, description: بالثواني / Seconds }
        user: { $ref: "#/components/schemas/User" }
...
    ResetPasswordRequest:
      type: object
      required: [email, token, password, password_confirmation]
      properties:
        email: { type: string, format: email }
        token: { type: string, description: رمز إعادة التعيين من البريد — صالح ساعة / Reset token from email — 1h validity }
        password: { type: string, format: password, minLength: 8 }
        password_confirmation: { type: string }

    # ---------------- Profile ----------------
...
    EvaluationStatusResponse:
      type: object
      required: [status]
      properties:
        status: { $ref: "#/components/schemas/EvaluationStatus" }
        evaluation_id: { type: [integer, "null"], format: int64 }
        overall_score: { type: [number, "null"] }
        message: { type: [string, "null"], description: رسالة الحالة (processing/failed) / Status message }
        started_at: { type: [string, "null"], format: date-time }
        eta_seconds: { type: [integer, "null"] }
        last_evaluation_at: { type: [string, "null"], format: date-time }
        next_evaluation_at: { type: [string, "null"], format: date-time }

    # ---------------- Search ----------------
...